More flexibility for implementors of persistence plugin
Operations is now a parent interface of transactions. This allow more
flexibility in commit strategies without having to duplicate default
writer logic.

Test now expect a PersistenceHandlers class rather than a store.
