+++ impl/persistence/api/src/main/java/io/serverlessworkflow/impl/persistence/AbstractPersistenceInstanceReader.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2020-Present The Serverless Workflow Specification Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.serverlessworkflow.impl.persistence;
+
+import io.serverlessworkflow.impl.WorkflowDefinition;
+import io.serverlessworkflow.impl.WorkflowInstance;
+import java.util.Optional;
+import java.util.stream.Stream;
+
+public abstract class AbstractPersistenceInstanceReader implements PersistenceInstanceReader {
+
+  protected final Stream<WorkflowInstance> scanAll(
+      PersistenceInstanceOperations operations,
+      WorkflowDefinition definition,
+      String applicationId) {
+    return operations
+        .scanAll(applicationId, definition)
+        .map(v -> new WorkflowPersistenceInstance(definition, v));
+  }
+
+  protected final Optional<WorkflowInstance> find(
+      PersistenceInstanceOperations operations, WorkflowDefinition definition, String instanceId) {
+    return operations
+        .readWorkflowInfo(definition, instanceId)
+        .map(i -> new WorkflowPersistenceInstance(definition, i));
+  }
+
+  @Override
+  public void close() {}
+}
